+++ CHANGELOG.md
@@ -39,5 +39,9 @@
   * Change: Now requires all views to be explicitly defined (and will not automatically use the underlying member view if it exists). To define a view for member element (wrapping it in a collection) one can use the new member_view.
 * Application.setup now returns the application instance so it can be chained to other application methods.
   * This allows a more compact way to rackup the application: `run Praxis::Application.instance.setup`.
+* Request headers are now backed by `Attributor::Hash` rather than (Structs)
+** Header keys are case sensitive now (although Praxis gracefully allows loading the uppercased RACK names )
+* Added tests for using `Attributor::Hash` attributes that can accept "extra" keys:
+** See the `bulk_create` action for an example applied to `Praxis::Multipart` (which derives from `Attributor::Hash`) payload [here](https://github.com/rightscale/praxis/blob/master/spec/spec_app/design/resources/instances.rb).
 	
 ## 0.9 Initial release